requirements.txt
asgiref==3.7.2
contourpy==1.2.0
cycler==0.12.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fonttools==4.44.0
gunicorn==23.0.0
imageio==2.32.0
importlib-resources==6.1.1
inflection==0.5.1
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
pygame==2.5.2
PyJWT==2.9.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
scikit-image==0.22.0
scipy==1.11.3
setuptools==72.1.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tifffile==2023.9.26
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
zipp==3.17.0
